requirements.txt
altair==5.5.0
attrs==25.1.0
blinker==1.9.0
branca==0.8.1
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
folium==0.20.0
gitdb==4.0.12
GitPython==3.1.44
haversine==2.9.0
idna==3.10
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
narwhals==1.29.1
numpy==2.2.3
packaging==24.2
pandas==2.2.3
patsy==1.0.1
pillow==11.1.0
plotly==6.2.0
plotly-express==0.4.1
protobuf==5.29.3
pyarrow==19.0.1
pydeck==0.9.1
pygame==2.6.1
python-dateutil==2.9.0.post0
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rpds-py==0.23.1
scipy==1.16.0
six==1.17.0
smmap==5.0.2
statsmodels==0.14.5
streamlit==1.43.1
streamlit_folium==0.25.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
xyzservices==2025.4.0